+++ CHANGELOG.markdown
@@ -1,7 +1,11 @@
 0.15.1
 ------
-* Bump support for `cereal`.
-* Drop `Trustworthy` claim, in `Data.Bytes.Put` as we now can sometimes infer `Safe`.
+* Drop `Trustworthy` claim in `Data.Bytes.Put` as we now can sometimes infer `Safe`.
+
+0.15.0.2
+--------
+* Bump `cereal` bound for 0.5.0.0
+* Add instance `Serial Natural`
 
 0.15.0.1
 --------